requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.20.0
python-multipart==0.0.6
python-dotenv==0.20.0
pydantic==2.10.3
aiofiles==23.2.1
psutil==5.9.5

# ML and Processing
torch==2.0.1
spacy==3.8.3
gliner==0.2.1
numpy==1.24.3
pandas==2.0.3

# Database
clickhouse-connect==0.6.8
asyncio==3.4.3

# Monitoring
prometheus-client==0.17.1
python-json-logger==2.0.7

# File processing
python-magic==0.4.27
pdf2image==1.16.3
pytesseract==0.3.10
pillow==10.0.0

# Type stubs
types-aiofiles==23.2.0.0
types-psutil==5.9.5.20240106
types-prometheus-client==0.17.0.4
types-PyYAML==6.0.12.12
types-requests==2.31.0.2
types-setuptools==69.1.0.0
types-urllib3==1.26.25.14
types-python-dateutil==2.8.19.20240106
types-tabulate==0.9.0.20240106

# Development tools
mypy==1.8.0
mypy-extensions==1.0.0
black==23.3.0
isort==5.12.0
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.1.0

# Error handling and retry
tenacity==8.2.3

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Optional: Development tools
ipython==8.12.0
jupyter==1.0.0